Bug 913120 Move places tests to its own addon to attempt to fix intermittent failures
