+++ test/addons/places/favicon-helpers.js
