+++ test/addons/places/package.json
@@ -1,0 +1,3 @@
+{
+  "id": "test-places"
+}
