+++ test/addons/places/tests/test-places-host.js
@@ -23,7 +23,7 @@ require('sdk/places/host/host-query');
 const {
   invalidResolve, invalidReject, createTree,
   compareWithHost, createBookmark, createBookmarkTree, resetPlaces
-} = require('./places-helper');
+} = require('../places-helper');
 
 const bmsrv = Cc['@mozilla.org/browser/nav-bookmarks-service;1'].
                     getService(Ci.nsINavBookmarksService);
@@ -301,5 +301,3 @@ exports.testGetAllChildren = function (assert, done) {
 
 before(exports, (name, assert, done) => resetPlaces(done));
 after(exports, (name, assert, done) => resetPlaces(done));
-
-require('test').run(exports);
